Change some -[NSApp beginSheet:…] to -[NSWindow beginSheet:…]. More pokes.

+++ Pfhorge Source/View and Controller/Doc Controllers/LELevelWindowController.h
@@ -81,7 +81,7 @@ extern NSNotificationName const PhLevelDidChangeNameNotification;
     
     BOOL closeIfLevelSettingsCancled;
     
-    int currentLevelLoaded;
+    NSInteger currentLevelLoaded;
     
     PhAnnotationNote *tmpNote;
 }
